python/Windows/shared/shared_modules/PiXYZAPI-2024.2.0.6/PiXYZAPI-2024.2.0.6-win64/PixyzUI_Sources/requirements.txt
pyinstaller
imgui-bundle==1.0
pyopengl
pillow
pynvml
psutil
jedi
glfw
